requirements.txt
nicegui>=2.0
yt-dlp>=2024.0
google-genai>=1.0
openai>=1.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
pyyaml>=6.0
python-dotenv>=1.0
aiofiles>=24.0
